+++ src/components/AboutUsCa.vue
@@ -5,20 +5,48 @@
       anat creixent amb el temps. Als seus inicis la reparació d'automòbils es basava principalment
       en la destresa manual i el enginy. Actualment els cotxes han evolucionat fins a l'extrem de
       portar incorporats ordinadors de control.
-    </p>
+    </p><br />
     <p>
       Quan Teraac va iniciar la seva tasca de servei ho va fer en un local petit i amb l'Andreu com
       a unic treballador, eren temps dificils i de molta feina.
-    </p>
+    </p><br />
     <p>
       Mes tard es van anar incorporant treballadors i l'empresa es va dotar d'unes instal·lacions
       més amplies i mes adequades al nou volum de treball que l'empresa havia de servir.
-    </p>
+    </p><br />
     <p>
       Actualment Teraac es caracteritza per ser una empresa que aquesta proveïda dels darrers
       equipaments en matèria de diagnòstic i reparació de les avaries de l'automòbil, amb un
       personal altament qualificat que continua formacio, sempre a disposicio i per donar el millor
       servei al client.
-    </p>
+    </p><br />
+  </v-card-text>
+  <v-card-title>
+    Bosch Car Service
+  </v-card-title>
+  <v-card-text>
+    <p>Els conductors ho tenen clar. Pensar en Bosch Car Service és garantir qualitat per al cotxe al millor preu,
+      sigui de la marca que sigui, en diagnosi, manteniment, reparacions, o en la revisió del cotxe nou sense perdre la
+      garantia dorigen. Bosch Car Service. Qualitat, Preu i tots els Serveis”.
+    </p><br />
+    <p>
+      La Xarxa oficial Bosch Car Service (BCS), amb més de 600 tallers repartits per tot Espanya, ofereix un servei
+      integral per a qualsevol marca d´automòbil. Perquè vostè exigeix per al seu cotxe un servei professional, de
+      qualitat,
+      confiança ia un preu raonable, els tallers Bosch Car Service (BCS) fan des del manteniment més senzill
+      fins a la reparació més complexa ja que compten amb els millors mecànics per garantir la reparació i
+      manteniment del vostre automòbil.
+    </p><br />
+    <p>
+      Els tallers Bosch Car Service (BCS) reben suport tècnic, empresarial i comercial per part de Bosch, líder
+      mundial com a fabricant de peces de l'automòbil.
+    </p><br />
+    <p>
+      Tot per al vostre automòbil: revisions, desgast d'elements, pre-itv, revisions de nivell, manteniments dins i
+      fora del període de garantia, electricitat i electrònica, mecànica en general, reparacions complexes, injecció
+      dièsel i benzina, frens, ABS…
+    </p><br />
+    <p>Xarxa Oficial de tallers Bosch Car Service.</p><br />
+    <p><b>Qualitat, Preu i tots els serveis</b></p><br />
   </v-card-text>
 </template>