+++ src/components/CookiesNotice.vue
@@ -2,11 +2,13 @@
   <v-app-bar class="cookies-notice-bar" flat height="40" v-if="cookieNotSet">
     <v-app-bar-title class="cookies-notice-title">
       <i18n-t keypath="legal.cookie-notice" scope="global">
-        <a href="#/legal-notice" target="_blank">{{ $t('legal.notice') }}</a>
+        <router-link @click.stop="closeCookiesNotice" to="legal-notice">
+          {{ $t('legal.title') }}
+        </router-link>
       </i18n-t>
     </v-app-bar-title>
     <template v-slot:append>
-      <v-btn @click="closeCookiesNotice" icon="mdi-close"></v-btn>
+      <v-btn @click.stop="closeCookiesNotice" icon="mdi-close"></v-btn>
     </template>
   </v-app-bar>
 </template>
@@ -45,8 +47,8 @@ a {
 }
 
 .cookies-notice-bar {
-  background-color: #ffcdd2 !important;
-  color: #e57373 !important;
+  background-color: #d1e4ff !important;
+  color: #56B0FF !important;
 }
 
 .cookies-notice-title {