+++ src/components/LegalCa.vue
@@ -1,295 +1,77 @@
 <template>
-  <v-card-subtitle>1.- Avís Legal</v-card-subtitle>
-  <v-card-text
-    >En compliment de la Llei 34/2002, de serveis de la societat de la informació i de comerç
-    Electrònic (LSSI), us informem que:</v-card-text
-  >
-  <v-card-subtitle>1.1- Dades identificatives</v-card-subtitle>
-  <v-card-text>
-    <p>El lloc web www.teraac.com (d'ara endavant, el “Lloc Web”) és titularitat de:</p>
-    <p>* Empresa: Teraac SL</p>
-    <p>* CIF: B61267654</p>
-    <p>* Registre mercantil:</p>
-    <p>* Domicili social: Carrer Pompeu Fabra, 50, 08922 Santa Coloma de Gramenet</p>
-    <p>* Telèfon: 93.386.12.68</p>
-    <p>* Correu electrònic: info@teraac.com</p>
-  </v-card-text>
-  <v-card-subtitle>1.2.- Condicions d'ús.</v-card-subtitle>
-  <v-card-text>
-    Aquest Lloc Web tenen per objecte facilitar als visitants (d'ara endavant, “l'Usuari”), el
-    coneixement de les activitats que aquesta empresa realitza i dels serveis que presta. L´ús
-    d´aquest lloc web implica l´acceptació expressa de les condicions aquí exposades, sense
-    perjudici d'aquelles particulars que poguessin aplicar-se a alguns dels serveis concrets oferts
-    a través del lloc web. L'usuari es compromet a utilitzar el Lloc Web i els seus continguts, amb
-    fins lícits que no lesionin els interessos del Titular drets de tercers.
-  </v-card-text>
-  <v-card-subtitle>1.3.- Propietat intel·lectual i industrial</v-card-subtitle>
-  <v-card-text>
-    <p>
-      El disseny del portal i els seus codis font, així com els logos, textos, marques i altres
-      signes distintius que apareixen al mateix pertanyen al Titular del lloc web i estan protegits
-      pels corresponents drets de propietat intel·lectual i industrial.
-    </p>
-    <p>
-      Queda expressament prohibida la reproducció, distribució o la comunicació pública, inclosa la
-      seva modalitat de posada a disposició, de la totalitat o part dels continguts d'aquesta pàgina
-      Web, amb qualsevol fi, a qualsevol suport i per qualsevol mitjà tècnic, sense lautorització
-      del Titular. No obstant això, lusuari pot visualitzar tots els elements, imprimir-los,
-      copiar-los i emmagatzemar-los al disc dur del vostre ordinador oa qualsevol altre suport físic
-      sempre que sigui, única i exclusivament, per a ús personal i privat.
-    </p>
-  </v-card-text>
-  <v-card-subtitle>1.4.- Enllaços de Tercers</v-card-subtitle>
-  <v-card-text>
-    <p>
-      El Titular no es fa responsable de la legalitat d'altres llocs web de tercers des dels quals
-      es pugui accedir al portal ni per la legalitat d'altres llocs web de tercers, que poguessin
-      estar vinculats o enllaçats des de aquest portal.
-    </p>
-    <p>
-      Així mateix, El Titular no es fa responsable de l'ús que tercers facin de la informació
-      publicada al portal, ni tampoc dels danys soferts o pèrdues econòmiques que, de forma directa
-      o indirecta, produeixin o puguin produir perjudicis econòmics, materials o sobre dades,
-      provocats per l'ús de la informació esmentada.
-    </p>
-  </v-card-text>
-  <v-card-subtitle>1.5.- Modificació de les Condicions</v-card-subtitle>
-  <v-card-text>
-    <p>
-      El Titular es reserva el dret d'efectuar, en qualsevol moment i sense necessitat de previ
-      avís, modificacions i actualitzacions de la informació continguda al Lloc Web Web oa la
-      configuració i presentació d'aquesta.
-    </p>
-  </v-card-text>
-  <v-card-subtitle>1.6.- Llei Aplicable i Jurisdicció</v-card-subtitle>
-  <v-card-text>
-    <p>
-      Totes les controvèrsies que se suscitin entre el Titular i l'Usuari relatives a l'Avís Legal
-      es regiran pels seus pròpies clàusules i, d'acord amb la legislació espanyola, sotmetent-se
-      expressament les parts a la jurisdicció dels Jutjats i Tribunals del domicili del Titular.
-    </p>
-  </v-card-text>
-  <v-card-subtitle>1.7.- Protecció de dades personals</v-card-subtitle>
-  <v-card-text>
-    <p>
-      Per utilitzar alguns dels serveis al Lloc Web, és possible que l'Usuari hagi de proporcionar
-      prèviament certes dades de caràcter personal mitjançant formularis o altres mitjans. Per això,
-      el Titular us informa que tractarà les dades personals d'acord amb les normatives vigents de
-      protecció de dades personals. Per saber-ne més, consulta la Política de Privadesa.
-    </p>
-  </v-card-text>
-  <v-card-subtitle>2.0.- Política de galetes</v-card-subtitle>
-  <v-card-text>
-    <p>
-      El lloc web de www. teraac.com, utilitza cookies per al correcte funcionament i visualització
-      del lloc web per part de lusuari, proporcionar una millor experiència en la navegació, així
-      com la recollida destadístiques.
-    </p>
-  </v-card-text>
-  <v-card-subtitle>2.1.- Què és una galeta i per a què serveix?</v-card-subtitle>
-  <v-card-text>
-    <p>
-      Les cookies són petits fitxers d'informació que el servidor envia a l'ordinador o dispositiu
-      de qui accedeix a la pàgina web, i ajuden a millorar la qualitat de la nostra web, permetent
-      controlar quines pàgines són útils, quins no i quinsón susceptibles de millora. Entre altres
-      coses, permeten emmagatzemar i recuperar informació sobre els hàbits de navegació d'un usuari
-      o del vostre equip, veure quines pàgines es visiten, el temps de connexió, analitzar el
-      funcionament d'una pàgina, si hi ha problemes tècnics, etc. Depenent de la informació que
-      continguin i la manera com s'utilitzi el terminal, es poden utilitzar per reconèixer l'usuari.
-    </p>
-    <p>
-      Les galetes són essencials per al funcionament d'internet, aportant innombrables avantatges en
-      la prestació de serveis interactius, facilitant-te la navegació i usabilitat de la nostra web.
-      En cap cas les galetes podrien fer malbé el teu equip. Per contra, el que estiguin actives ens
-      ajuda a identificar i resoldre els errors. El lloc web és accessible sense necessitat que les
-      galetes estiguin activades, si bé, la seva desactivació pot impedir el correcte funcionament
-      del mateix.
-    </p>
-  </v-card-text>
-  <v-card-subtitle>2.2.- Tipus de galetes</v-card-subtitle>
-  <v-card-text>
-    <p>Hi ha diversos tipus de cookies en funció de diferents criteris.</p>
-    <p>En funció del temps que romanen activades:</p>
-    <p>
-      * Galetes de sessió. Són un tipus de cookies dissenyades per demanar i emmagatzemar dades
-      mentre l'usuari accedeix a una pàgina web.
-    </p>
-    <p>
-      * Cookies permanents: Són un tipus de cookies en què les dades segueixen emmagatzemades al
-      terminal i poden ser accedits i tractats durant un període definit pel responsable de la
-      galeta.
-    </p>
-    <p>En funció de la titularitat:</p>
-    <p>* Cookies pròpies. Són les que pertanyen al titular de la pàgina web.</p>
-    <p>
-      * Galetes de tercers. Són aquelles la titularitat de les quals és d'un tercer, diferent del
-      titular de la pàgina web, que serà qui tracti la informació recollida.
-    </p>
-    <p>En funció de la seva finalitat:</p>
-    <p>
-      * Cookies tècniques i/o de personalització. Són aquelles que serveixen per millorar el servei,
-      localitzar incidències, reconèixer l'usuari, etc.
-    </p>
-    <p>
-      * Galetes d'anàlisi i/o de publicitat. Són aquelles que serveixen per analitzar informació
-      sobre la navegació i oferir publicitat, sigui genèrica o personalitzada.
-    </p>
-  </v-card-text>
-  <v-card-subtitle>2.3.- Cookies utilitzades al lloc web</v-card-subtitle>
-  <v-card-text>
-    <p>
-      En aquest lloc web, en concret, s'utilitzen cookies pròpies i de tercers, de tipus tècnic,
-      analític i de publicitat. En detall, serien les cookies següents:
-    </p>
-    <p>* [detall cookies utilitzades]</p>
-    <p>
-      Aquestes eines no obtenen dades personals dels usuaris ni de l'adreça postal des d'on es
-      connecten. La informació que obtenen està relacionada amb, el nombre de pàgines visitades,
-      l'idioma, la xarxa social on es publiquen les nostres notícies, la ciutat a la qual està
-      assignada l'adreça IP des de la qual accedeixen els usuaris, el nombre d'usuaris que ens
-      visiten, la freqüència i reincidència de les visites, el temps de visita, el navegador que
-      usen, l'operador o tipus de terminal des del qual es fa la visita.
-    </p>
-  </v-card-text>
-  <v-card-subtitle>2.4.- Consentiment</v-card-subtitle>
-  <v-card-text>
-    <p>
-      Acceptant les galetes, l'usuari consent expressament l'ús de galetes per part del web, segons
-      la descripció que es detalla a continuació, excepte en la mesura que hagi modificat la
-      configuració de la seva navegador per rebutjar la utilització de les mateixes. No cal el
-      consentiment per a la instal·lació de cookies tècniques o les que siguin estrictament
-      necessàries per a la prestació d'un servei de la societat de la informació expressament
-      sol·licitat pel destinatari. Sentendrà que satorgat si lusuari. segueix utilitzant la pàgina
-      web, sense perjudici que en qualsevol moment podrà revocar-ho i desactivar les cookies.
-    </p>
-  </v-card-text>
-  <v-card-subtitle>2.5.- Deshabilitar les galetes</v-card-subtitle>
-  <v-card-text>
-    <p>
-      L'usuari pot permetre, bloquejar o eliminar les cookies instal·lades al vostre equip
-      mitjançant la configuració de les opcions del vostre navegador d'Internet. En cas que no
-      permeti la instal·lació de cookies al teu navegador és possible que no es pugui accedir a
-      alguns dels serveis i que l'experiència a la nostra web pugui resultar menys satisfactòria.
-      Als següents enllaços podeu trobar més informació per configurar els diferents navegadors: –
-      Google Chrome – Mozilla Firefox – Internet Explorer – Safari
-    </p>
-  </v-card-text>
-  <v-card-subtitle>3.0.- Política de Privadesa</v-card-subtitle>
-  <v-card-text>
-    <p>
-      Les dades de caràcter personal que es poguessin demanar directament de l'interessat (l'Usuari)
-      seran tractades de forma confidencial i quedaran incorporats a la corresponent activitat de
-      tractament titularitat del Responsable del Tractament, d'acord amb les normatives vigents de
-      protecció de dades personalshttps://www.aepd.es/normativa/index.html
-    </p>
-    <p>
-      En aquest sentit, tractarem les dades personals amb lleialtat, transparència i licitud,
-      sol·licitant la teva consentiment, tractant les dades estrictament necessàries per a les
-      finalitats requerides, limitant-ne el seu conservació als terminis legalment establerts o els
-      necessaris per a les finalitats establertes. així mateix, hem aplicat les mesures tècniques i
-      organitzatives adequades, per garantir la seguretat i confidencialitat de les dades,
-      evitant-ne l'ús indegut, destrucció no autoritzada o accessos no autoritzats.
-    </p>
-  </v-card-text>
-  <v-card-subtitle>3.1.- Responsable del Tractament</v-card-subtitle>
-  <v-card-text>
-    <p>* Empresa: Teraac SL</p>
-    <p>* CIF: B61267654</p>
-    <p>* Domicili social: Carrer Pompeu Fabra, 50, 08922 Santa Coloma de Gramenet</p>
-    <p>* Telèfon: 93.386.12.68</p>
-    <p>* Correu electrònic: info@teraac.com</p>
-  </v-card-text>
-  <v-card-subtitle>3.2.- Finalitats del tractament</v-card-subtitle>
-  <v-card-text>
-    <p>
-      Les dades personals sol·licitades són necessàries per gestionar els serveis que el Responsable
-      del Tractament ofereix al Lloc Web, per la qual cosa, si no ens els facilita, no podrem
-      atendre'l correctament ni prestar-li el servei que ha sol·licitat. En aquest Lloc Web, tractem
-      les vostres dades personals amb les següents finalitats:
-    </p>
-    <p>Formulari de contacte</p>
-    <p>
-      Per recollir informació sobre consultes dels Usuaris. Aquest tractament es legitima per
-      l'interès del interessat. En aquest formulari se sol·liciten les dades personals següents:
-    </p>
-    <p>* Nom, idioma, correu electrònic</p>
-    <p>
-      Aquestes dades no seran cedides a tercers, i es conservaran mentre hi hagi una activitat
-      comunicativa i, llevat que l'Usuari sol·liciti la baixa, moment en què quedaran bloquejats per
-      poder demostrar que l'usuari va consentir el tractament.
-    </p>
-    <p>Ordres de reparació (OR)</p>
-    <p>
-      A les OR es recullen les dades necessàries per prestar els serveis del taller. No
-      facilitar-los pot impedir que no es presti el servei adequadament. Aquest tractament es
-      legitima per l'execució del contracte i d'una obligació legal. En aquest formulari se
-      sol·liciten les dades personals següents:
-    </p>
-    <p>* Nom i cognoms, DNI, telèfon, correu electrònic, matrícula vehicle, adreça</p>
-    <p>
-      Aquestes dades no seran cedides a tercers llevat que sigui necessari per prestar els nostres
-      serveis, i es conservaran pel temps necessari per al compliment de les finalitats previstes, o
-      mentre hi hagi obligacions legals.
-    </p>
-  </v-card-text>
-  <v-card-subtitle>3.3.- Cessió de dades</v-card-subtitle>
-  <v-card-text>
-    <p>
-      El Responsable del Tractament no cedirà les dades a tercers llevat del consentiment de
-      l'Usuari, llevat que existeixi una obligació legal, o un interès per part de l'usuari.
-    </p>
-  </v-card-text>
-  <v-card-subtitle>3.4.- Obtenció de les dades</v-card-subtitle>
-  <v-card-text>
-    <p>
-      Les dades personals són obtingudes mitjançant el consentiment de l'Usuari, mitjançant
-      formularis o altres mètodes transparents i no invasius amb la privadesa de l'Usuari. Aquest és
-      el responsable de l'exactitud i veracitat de les dades facilitades, i de la seva correcció en
-      cas que hagin estat actualitzades. Els menors de 14 anys requeriran l'autorització dels pares
-      o d'un representant legal per facilitar les vostres dades personals.
-    </p>
-  </v-card-text>
-  <v-card-subtitle>3.5.- Drets dels Usuaris</v-card-subtitle>
-  <v-card-text>
-    <p>
-      Qualsevol Usuari pot retirar el seu consentiment en qualsevol moment, quan aquest s'hagi
-      atorgat per al tractament de les vostres dades.
-    </p>
-    <p>Dret d'accés</p>
-    <p>
-      L'Usuari té el dret a conèixer quina informació disposa el Responsable del Tractament ia
-      obtenir-ne una còpia dels mateixos.
-    </p>
-    <p>Dret de Rectificació</p>
-    <p>
-      L'Usuari té el dret que es rectifiquin les dades per actualitzar-les o perquè són inexactes.
-    </p>
-    <p>Dret de Limitació i Oposició</p>
-    <p>
-      L'Usuari té el dret a limitar i oposar-se al tractament en determinades circumstàncies
-      reconegudes de forma expressa a la normativa.
-    </p>
-    <p>Dret de Supressió</p>
-    <p>
-      L'Usuari té el dret que se suprimeixin les dades. Quan la conservació sigui necessària per al
-      compliment d'una obligació legal, per desenvolupar una relació contractual, o per determinar
-      possibles responsabilitats del Responsable del Tractament i de l'Usuari, les dades romandran
-      bloquejades de manera que no es pugui fer tractament sobre aquests.
-    </p>
-    <p>
-      Podeu exercitar aquests drets mitjançant l'enviament d'un correu electrònic a: info@teraac.com
-    </p>
-    <p>
-      No obstant, també us podeu adreçar davant l'Autoritat de Protecció de Dades Personals
-      (www.aepd.es)
-    </p>
-  </v-card-text>
-  <v-card-subtitle>3.6.- Modificació de la Política de Privadesa</v-card-subtitle>
-  <v-card-text>
-    <p>
-      El Responsable del Tractament es reserva el dret d'efectuar, en qualsevol moment i sense
-      necessitat previ avís, modificacions i actualitzacions del contingut d'aquesta Política de
-      Privadesa”
-    </p>
-  </v-card-text>
-</template>
+  <v-card-text>Aquest lloc web és propietat de TERAAC SL i es fa servir per a proporcionar informació als usuaris sobre
+    els productes i serveis que oferim.</v-card-text>
+  <v-card-subtitle>Propietat intel·lectual</v-card-subtitle>
+  <v-card-text>Tots els continguts d'aquest lloc web, incloent-hi textos, imatges, gràfics, logotips i marques comercials,
+    són propietat de TERAAC SL o dels seus respectius propietaris i estan protegits per les lleis de propietat
+    intel·lectual. No es permet la reproducció, distribució, modificació o explotació de cap contingut d'aquest lloc sense
+    el consentiment previ per escrit de TERAAC SL.</v-card-text>
+  <v-card-subtitle>Limitació de responsabilitat</v-card-subtitle>
+  <v-card-text>Aquest lloc web proporciona informació general sobre els productes i serveis que oferim. Tot i que ens
+    esforcem per proporcionar informació precisa i actualitzada, no garantim la exactitud, integritat o adequació de la
+    informació proporcionada en aquest lloc web.</v-card-text>
+  <v-card-text>En accedir i utilitzar aquest lloc web, l'usuari accepta que TERAAC SL no serà responsable de qualsevol
+    pèrdua o dany directe, indirecte, incidental, especial o conseqüencial que pugui resultar de l'ús o la impossibilitat
+    d'usar aquest lloc web.</v-card-text>
+  <v-card-subtitle>Enllaços a tercers</v-card-subtitle>
+  <v-card-text>Aquest lloc web pot contenir enllaços a altres llocs web que no són propietat de TERAAC SL. No tenim cap
+    control sobre aquests llocs web i no som responsables del contingut o les pràctiques de privacitat d'aquests llocs
+    web.</v-card-text>
+  <v-card-subtitle>Política de privacitat</v-card-subtitle>
+  <v-card-text>La nostra política de privacitat descriu com recopilem, utilitzem i compartim informació recopilada a
+    través d'aquest lloc web. En accedir i utilitzar aquest lloc web, l'usuari accepta la nostra política de
+    privacitat.</v-card-text>
+  <v-card-subtitle>Canvis a l'avís legal</v-card-subtitle>
+  <v-card-text>Ens reservem el dret de modificar aquest avís legal en qualsevol moment. Es recomana als usuaris que
+    revisin periòdicament aquest avís legal per a estar informat dels canvis.</v-card-text>
+  <v-card-subtitle>Avís legal de l'ús de cookies</v-card-subtitle>
+  <v-card-text>Aquest lloc web utilitza cookies per millorar l'experiència de l'usuari i proporcionar informació analítica
+    per a millorar el lloc. Les cookies són petits fitxers de dades que es descarreguen al seu dispositiu quan accedeix al
+    lloc web. Aquestes cookies no recopilen informació personal identifiable de l'usuari.</v-card-text>
+  <v-card-text>Les cookies que utilitzem són cookies d'anàlisi que ens permeten recopilar informació anònima sobre com els
+    usuaris naveguen pel nostre lloc web. Això ens ajuda a millorar el contingut i la funcionalitat del lloc per fer-lo
+    més útil per a l'usuari.</v-card-text>
+  <v-card-text>En utilitzar aquest lloc web, vostè està acceptant l'ús de cookies. Si no desitja acceptar l'ús de cookies,
+    pot modificar la configuració del navegador per a bloquejar-les. No obstant això, si bloqueja les cookies, és possible
+    que no pugui accedir a determinades parts del lloc web o que la seva experiència de navegació sigui menys
+    satisfactòria.</v-card-text>
+  <v-card-text>Aquest lloc web també utilitza serveis de tercers que poden utilitzar cookies. És possible que aquests
+    tercers recullin informació sobre l'activitat de l'usuari en aquest lloc web i altres llocs web per a proporcionar
+    publicitat personalitzada basada en els interessos de l'usuari.</v-card-text>
+<v-card-text>Per a més informació sobre l'ús de cookies en aquest lloc web, si us plau, consulteu la nostra política de
+  privacitat a continuació.</v-card-text>
+<v-card-title>Política de privacitat</v-card-title>
+<v-card-text>La privacitat dels nostres usuaris és important per a nosaltres i ens comprometem a protegir-la. Aquesta
+  política de privacitat descriu com recopilem, utilitzem i compartim informació recopilada a través d'aquest lloc
+  web.</v-card-text>
+<v-card-subtitle>Recopilació d'informació</v-card-subtitle>
+<v-card-text>Podem recopilar informació personal identificable, com ara nom, adreça de correu electrònic, número de
+  telèfon, adreça postal, etc. només si és proporcionada voluntàriament pels usuaris a través dels formularis de
+  contacte o subscripció al nostre newsletter.</v-card-text>
+<v-card-text>A més, a través de l'ús de cookies, recopilem informació no personal identifiable com ara l'adreça IP, el
+  tipus de navegador, el proveïdor d'Internet, la pàgina de referència i la sortida, i altres dades estadístiques
+  relacionades amb l'ús del lloc web.</v-card-text>
+<v-card-subtitle>Ús de la informació</v-card-subtitle>
+<v-card-text>La informació recopilada a través del lloc web es fa servir per millorar la funcionalitat i el contingut
+  del lloc i per proporcionar als usuaris una millor experiència de navegació. També podem utilitzar la informació per a
+  respondre a les consultes dels usuaris, per a enviar el newsletter sol·licitat o per a altres finalitats comunicades
+  als usuaris en el moment de recopilar la informació.</v-card-text>
+<v-card-subtitle>Compartició de la informació</v-card-subtitle>
+<v-card-text>No compartim ni venem informació personal identifiable a tercers sense el consentiment dels usuaris,
+  excepte en els casos en què estiguem obligats per llei o per una ordre judicial.</v-card-text>
+<v-card-text>No obstant això, podem compartir informació no personal identifiable amb tercers, com ara proveïdors de
+  serveis d'anàlisi web, amb l'objectiu de millorar el lloc web i proporcionar als usuaris una millor experiència de
+  navegació.</v-card-text>
+<v-card-subtitle>Seguretat</v-card-subtitle>
+<v-card-text>Ens comprometem a protegir la informació dels usuaris i hem implementat mesures de seguretat tècniques i
+  organitzatives per protegir la informació recopilada a través del lloc web.</v-card-text>
+<v-card-subtitle>Drets dels usuaris</v-card-subtitle>
+<v-card-text>Els usuaris tenen dret a accedir, corregir, actualitzar i esborrar la informació personal que ens hagin
+  proporcionat. Si us plau, poseu-vos en contacte amb nosaltres a través de les dades de contacte proporcionades al lloc
+  web per exercir aquests drets.</v-card-text>
+<v-card-subtitle>Canvis a la política de privacitat</v-card-subtitle>
+<v-card-text>Ens reservem el dret de modificar aquesta política de privacitat en qualsevol moment. Es recomana als
+  usuaris que revisin periòdicament aquesta política de privacitat per a estar informat dels canvis.</v-card-text>
+<v-card-subtitle>Contacte</v-card-subtitle>
+<v-card-text>Si teniu qualsevol pregunta o comentari sobre aquest avís legal, si us plau, poseu-vos en contacte amb
+  nosaltres a través dels detalls de contacte proporcionats al lloc web.</v-card-text></template>